Refactor to change all references to Duke to Shirmin
