+++ docs/Duke/Duke.html
@@ -2,11 +2,11 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (19) on Tue Jan 30 22:38:03 SGT 2024 -->
-<title>Duke</title>
+<title>Shirmin</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2024-01-30">
-<meta name="description" content="declaration: package: Duke, class: Duke">
+<meta name="description" content="declaration: package: Shirmin, class: Shirmin">
 <meta name="generator" content="javadoc/ClassWriterImpl">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.min.css" title="Style">
@@ -82,19 +82,19 @@ loadScripts(document, 'script');</script>
 <main role="main">
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="sub-title"><span class="package-label-in-type">Package</span>&nbsp;<a href="package-summary.html">Duke</a></div>
-<h1 title="Class Duke" class="title">Class Duke</h1>
+<div class="sub-title"><span class="package-label-in-type">Package</span>&nbsp;<a href="package-summary.html">Shirmin</a></div>
+<h1 title="Class Shirmin" class="title">Class Shirmin</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/19/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
-<div class="inheritance">Duke.Duke</div>
+<div class="inheritance">Shirmin.Duke</div>
 </div>
 <section class="class-description" id="class-description">
 <hr>
-<div class="type-signature"><span class="modifiers">public class </span><span class="element-name type-name-label">Duke</span>
+<div class="type-signature"><span class="modifiers">public class </span><span class="element-name type-name-label">Shirmin</span>
 <span class="extends-implements">extends <a href="https://docs.oracle.com/en/java/javase/19/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">Object</a></span></div>
-<div class="block">Represents the main class of the Duke application.
+<div class="block">Represents the main class of the Shirmin application.
  <p>
- This class is responsible for initializing and running the Duke application.
+ This class is responsible for initializing and running the Shirmin application.
  It integrates various classes in an OOP structure - Storage, TaskList, Parser, and Ui.</div>
 </section>
 <section class="summary">
@@ -107,9 +107,9 @@ loadScripts(document, 'script');</script>
 <div class="summary-table two-column-summary">
 <div class="table-header col-first">Constructor</div>
 <div class="table-header col-last">Description</div>
-<div class="col-constructor-name even-row-color"><code><a href="#%3Cinit%3E()" class="member-name-link">Duke</a>()</code></div>
+<div class="col-constructor-name even-row-color"><code><a href="#%3Cinit%3E()" class="member-name-link">Shirmin</a>()</code></div>
 <div class="col-last even-row-color">
-<div class="block">Initializes a new Duke application.</div>
+<div class="block">Initializes a new Shirmin application.</div>
 </div>
 </div>
 </section>
@@ -128,12 +128,12 @@ loadScripts(document, 'script');</script>
 <div class="col-first even-row-color method-summary-table method-summary-table-tab1 method-summary-table-tab4"><code>static void</code></div>
 <div class="col-second even-row-color method-summary-table method-summary-table-tab1 method-summary-table-tab4"><code><a href="#main(java.lang.String%5B%5D)" class="member-name-link">main</a><wbr>(<a href="https://docs.oracle.com/en/java/javase/19/docs/api/java.base/java/lang/String.html" title="class or interface in java.lang" class="external-link">String</a>[]&nbsp;args)</code></div>
 <div class="col-last even-row-color method-summary-table method-summary-table-tab1 method-summary-table-tab4">
-<div class="block">The entry point of the Duke application.</div>
+<div class="block">The entry point of the Shirmin application.</div>
 </div>
 <div class="col-first odd-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code>void</code></div>
 <div class="col-second odd-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code><a href="#run()" class="member-name-link">run</a>()</code></div>
 <div class="col-last odd-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4">
-<div class="block">Starts the execution of the Duke application.</div>
+<div class="block">Starts the execution of the Shirmin application.</div>
 </div>
 </div>
 </div>
@@ -154,11 +154,11 @@ loadScripts(document, 'script');</script>
 <ul class="member-list">
 <li>
 <section class="detail" id="&lt;init&gt;()">
-<h3>Duke</h3>
-<div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="element-name">Duke</span>()</div>
-<div class="block">Initializes a new Duke application.
+<h3>Shirmin</h3>
+<div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="element-name">Shirmin</span>()</div>
+<div class="block">Initializes a new Shirmin application.
  <p>
- This constructor initializes the Ui, Parser, Storage, and TaskList components of the Duke application.</div>
+ This constructor initializes the Ui, Parser, Storage, and TaskList components of the Shirmin application.</div>
 </section>
 </li>
 </ul>
@@ -173,7 +173,7 @@ loadScripts(document, 'script');</script>
 <section class="detail" id="run()">
 <h3>run</h3>
 <div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="return-type">void</span>&nbsp;<span class="element-name">run</span>()</div>
-<div class="block">Starts the execution of the Duke application.
+<div class="block">Starts the execution of the Shirmin application.
  <p>
  This method triggers the main functionality of the application,
  starting with the task list functionalities.</div>
@@ -183,9 +183,9 @@ loadScripts(document, 'script');</script>
 <section class="detail" id="main(java.lang.String[])">
 <h3>main</h3>
 <div class="member-signature"><span class="modifiers">public static</span>&nbsp;<span class="return-type">void</span>&nbsp;<span class="element-name">main</span><wbr><span class="parameters">(<a href="https://docs.oracle.com/en/java/javase/19/docs/api/java.base/java/lang/String.html" title="class or interface in java.lang" class="external-link">String</a>[]&nbsp;args)</span></div>
-<div class="block">The entry point of the Duke application.
+<div class="block">The entry point of the Shirmin application.
  <p>
- This main method initializes a new Duke instance and starts its execution.</div>
+ This main method initializes a new Shirmin instance and starts its execution.</div>
 <dl class="notes">
 <dt>Parameters:</dt>
 <dd><code>args</code> - The command-line arguments.</dd>
