+++ docs/Duke/Event.html
@@ -6,7 +6,7 @@
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2024-01-30">
-<meta name="description" content="declaration: package: Duke, class: Event">
+<meta name="description" content="declaration: package: Shirmin, class: Event">
 <meta name="generator" content="javadoc/ClassWriterImpl">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.min.css" title="Style">
@@ -82,18 +82,18 @@ loadScripts(document, 'script');</script>
 <main role="main">
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="sub-title"><span class="package-label-in-type">Package</span>&nbsp;<a href="package-summary.html">Duke</a></div>
+<div class="sub-title"><span class="package-label-in-type">Package</span>&nbsp;<a href="package-summary.html">Shirmin</a></div>
 <h1 title="Class Event" class="title">Class Event</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/19/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
-<div class="inheritance">Duke.Task
-<div class="inheritance">Duke.Event</div>
+<div class="inheritance">Shirmin.Task
+<div class="inheritance">Shirmin.Event</div>
 </div>
 </div>
 <section class="class-description" id="class-description">
 <hr>
 <div class="type-signature"><span class="modifiers">public class </span><span class="element-name type-name-label">Event</span>
-<span class="extends-implements">extends Duke.Task</span></div>
+<span class="extends-implements">extends Shirmin.Task</span></div>
 <div class="block">Represents an event, represented by a description string and two LocalDateTime objects - from and to</div>
 </section>
 <section class="summary">
@@ -115,7 +115,7 @@ loadScripts(document, 'script');</script>
 <div class="col-last odd-row-color">&nbsp;</div>
 </div>
 <div class="inherited-list">
-<h3 id="fields-inherited-from-class-Duke.Task">Fields inherited from class&nbsp;Duke.Task</h3>
+<h3 id="fields-inherited-from-class-Shirmin.Task">Fields inherited from class&nbsp;Shirmin.Task</h3>
 <code>description, isDone</code></div>
 </section>
 </li>
@@ -171,7 +171,7 @@ loadScripts(document, 'script');</script>
 </div>
 </div>
 <div class="inherited-list">
-<h3 id="methods-inherited-from-class-Duke.Task">Methods inherited from class&nbsp;Duke.Task</h3>
+<h3 id="methods-inherited-from-class-Shirmin.Task">Methods inherited from class&nbsp;Shirmin.Task</h3>
 <code>getDescription, getStatusIcon, isDone, markDone, markUndone</code></div>
 <div class="inherited-list">
 <h3 id="methods-inherited-from-class-java.lang.Object">Methods inherited from class&nbsp;java.lang.<a href="https://docs.oracle.com/en/java/javase/19/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">Object</a></h3>
@@ -284,7 +284,7 @@ loadScripts(document, 'script');</script>
  Example: "[E][✓] Meeting (from: Oct 12 2021 10:00 to: Oct 12 2021 11:00)"</div>
 <dl class="notes">
 <dt>Overrides:</dt>
-<dd><code>toString</code>&nbsp;in class&nbsp;<code>Duke.Task</code></dd>
+<dd><code>toString</code>&nbsp;in class&nbsp;<code>Shirmin.Task</code></dd>
 <dt>Returns:</dt>
 <dd>String representing the Event with its status, description, and time range.</dd>
 </dl>
