+++ docs/Duke/Parser.html
@@ -6,7 +6,7 @@
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2024-01-30">
-<meta name="description" content="declaration: package: Duke, class: Parser">
+<meta name="description" content="declaration: package: Shirmin, class: Parser">
 <meta name="generator" content="javadoc/ClassWriterImpl">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.min.css" title="Style">
@@ -82,11 +82,11 @@ loadScripts(document, 'script');</script>
 <main role="main">
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="sub-title"><span class="package-label-in-type">Package</span>&nbsp;<a href="package-summary.html">Duke</a></div>
+<div class="sub-title"><span class="package-label-in-type">Package</span>&nbsp;<a href="package-summary.html">Shirmin</a></div>
 <h1 title="Class Parser" class="title">Class Parser</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/19/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
-<div class="inheritance">Duke.Parser</div>
+<div class="inheritance">Shirmin.Parser</div>
 </div>
 <section class="class-description" id="class-description">
 <hr>
@@ -107,7 +107,7 @@ loadScripts(document, 'script');</script>
 <div class="summary-table two-column-summary">
 <div class="table-header col-first">Constructor</div>
 <div class="table-header col-last">Description</div>
-<div class="col-constructor-name even-row-color"><code><a href="#%3Cinit%3E(Duke.Ui)" class="member-name-link">Parser</a><wbr>(Duke.Ui&nbsp;ui)</code></div>
+<div class="col-constructor-name even-row-color"><code><a href="#%3Cinit%3E(Shirmin.Ui)" class="member-name-link">Parser</a><wbr>(Shirmin.Ui&nbsp;ui)</code></div>
 <div class="col-last even-row-color">
 <div class="block">Initializes a new Parser instance.</div>
 </div>
@@ -130,7 +130,7 @@ loadScripts(document, 'script');</script>
 <div class="col-last even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4">
 <div class="block">Parses a date string into a formatted date string.</div>
 </div>
-<div class="col-first odd-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code>Duke.Task</code></div>
+<div class="col-first odd-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code>Shirmin.Task</code></div>
 <div class="col-second odd-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code><a href="#parseTask(java.lang.String)" class="member-name-link">parseTask</a><wbr>(<a href="https://docs.oracle.com/en/java/javase/19/docs/api/java.base/java/lang/String.html" title="class or interface in java.lang" class="external-link">String</a>&nbsp;line)</code></div>
 <div class="col-last odd-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4">
 <div class="block">Parses a string into a Task object.</div>
@@ -153,9 +153,9 @@ loadScripts(document, 'script');</script>
 <h2>Constructor Details</h2>
 <ul class="member-list">
 <li>
-<section class="detail" id="&lt;init&gt;(Duke.Ui)">
+<section class="detail" id="&lt;init&gt;(Shirmin.Ui)">
 <h3>Parser</h3>
-<div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="element-name">Parser</span><wbr><span class="parameters">(Duke.Ui&nbsp;ui)</span></div>
+<div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="element-name">Parser</span><wbr><span class="parameters">(Shirmin.Ui&nbsp;ui)</span></div>
 <div class="block">Initializes a new Parser instance.</div>
 <dl class="notes">
 <dt>Parameters:</dt>
@@ -174,7 +174,7 @@ loadScripts(document, 'script');</script>
 <li>
 <section class="detail" id="parseTask(java.lang.String)">
 <h3>parseTask</h3>
-<div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="return-type">Duke.Task</span>&nbsp;<span class="element-name">parseTask</span><wbr><span class="parameters">(<a href="https://docs.oracle.com/en/java/javase/19/docs/api/java.base/java/lang/String.html" title="class or interface in java.lang" class="external-link">String</a>&nbsp;line)</span></div>
+<div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="return-type">Shirmin.Task</span>&nbsp;<span class="element-name">parseTask</span><wbr><span class="parameters">(<a href="https://docs.oracle.com/en/java/javase/19/docs/api/java.base/java/lang/String.html" title="class or interface in java.lang" class="external-link">String</a>&nbsp;line)</span></div>
 <div class="block">Parses a string into a Task object.
  <p>
  The method interprets a string representation of a task (e.g., from a saved file)
