+++ docs/Duke/package-summary.html
@@ -2,11 +2,11 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (19) on Tue Jan 30 22:38:03 SGT 2024 -->
-<title>Duke</title>
+<title>Shirmin</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2024-01-30">
-<meta name="description" content="declaration: package: Duke">
+<meta name="description" content="declaration: package: Shirmin">
 <meta name="generator" content="javadoc/PackageWriterImpl">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.min.css" title="Style">
@@ -65,10 +65,10 @@ loadScripts(document, 'script');</script>
 <div class="flex-content">
 <main role="main">
 <div class="header">
-<h1 title="Package Duke" class="title">Package Duke</h1>
+<h1 title="Package Shirmin" class="title">Package Shirmin</h1>
 </div>
 <hr>
-<div class="package-signature">package <span class="element-name">Duke</span></div>
+<div class="package-signature">package <span class="element-name">Shirmin</span></div>
 <section class="summary">
 <ul class="summary-list">
 <li>
@@ -77,27 +77,27 @@ loadScripts(document, 'script');</script>
 <div class="summary-table two-column-summary">
 <div class="table-header col-first">Class</div>
 <div class="table-header col-last">Description</div>
-<div class="col-first even-row-color class-summary class-summary-tab2"><a href="Deadline.html" title="class in Duke">Deadline</a></div>
+<div class="col-first even-row-color class-summary class-summary-tab2"><a href="Deadline.html" title="class in Shirmin">Deadline</a></div>
 <div class="col-last even-row-color class-summary class-summary-tab2">
 <div class="block">Represents a task with a deadline.</div>
 </div>
-<div class="col-first odd-row-color class-summary class-summary-tab2"><a href="Duke.html" title="class in Duke">Duke</a></div>
+<div class="col-first odd-row-color class-summary class-summary-tab2"><a href="Shirmin.html" title="class in Shirmin">Shirmin</a></div>
 <div class="col-last odd-row-color class-summary class-summary-tab2">
-<div class="block">Represents the main class of the Duke application.</div>
+<div class="block">Represents the main class of the Shirmin application.</div>
 </div>
-<div class="col-first even-row-color class-summary class-summary-tab2"><a href="Event.html" title="class in Duke">Event</a></div>
+<div class="col-first even-row-color class-summary class-summary-tab2"><a href="Event.html" title="class in Shirmin">Event</a></div>
 <div class="col-last even-row-color class-summary class-summary-tab2">
 <div class="block">Represents an event, represented by a description string and two LocalDateTime objects - from and to</div>
 </div>
-<div class="col-first odd-row-color class-summary class-summary-tab2"><a href="Parser.html" title="class in Duke">Parser</a></div>
+<div class="col-first odd-row-color class-summary class-summary-tab2"><a href="Parser.html" title="class in Shirmin">Parser</a></div>
 <div class="col-last odd-row-color class-summary class-summary-tab2">
 <div class="block">Represents a parser for interpreting user input and file data into Tasks.</div>
 </div>
-<div class="col-first even-row-color class-summary class-summary-tab2"><a href="Storage.html" title="class in Duke">Storage</a></div>
+<div class="col-first even-row-color class-summary class-summary-tab2"><a href="Storage.html" title="class in Shirmin">Storage</a></div>
 <div class="col-last even-row-color class-summary class-summary-tab2">
-<div class="block">Handles storage operations for Duke application tasks.</div>
+<div class="block">Handles storage operations for Shirmin application tasks.</div>
 </div>
-<div class="col-first odd-row-color class-summary class-summary-tab2"><a href="Todo.html" title="class in Duke">Todo</a></div>
+<div class="col-first odd-row-color class-summary class-summary-tab2"><a href="Todo.html" title="class in Shirmin">Todo</a></div>
 <div class="col-last odd-row-color class-summary class-summary-tab2">
 <div class="block">Represents a basic task without any additional date/time.</div>
 </div>
