+++ docs/Duke/package-tree.html
@@ -2,11 +2,11 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (19) on Tue Jan 30 22:38:03 SGT 2024 -->
-<title>Duke Class Hierarchy</title>
+<title>Shirmin Class Hierarchy</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <meta name="dc.created" content="2024-01-30">
-<meta name="description" content="tree: package: Duke">
+<meta name="description" content="tree: package: Shirmin">
 <meta name="generator" content="javadoc/PackageTreeWriter">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.min.css" title="Style">
@@ -48,21 +48,21 @@ loadScripts(document, 'script');</script>
 <div class="flex-content">
 <main role="main">
 <div class="header">
-<h1 class="title">Hierarchy For Package Duke</h1>
+<h1 class="title">Hierarchy For Package Shirmin</h1>
 </div>
 <section class="hierarchy">
 <h2 title="Class Hierarchy">Class Hierarchy</h2>
 <ul>
 <li class="circle">java.lang.<a href="https://docs.oracle.com/en/java/javase/19/docs/api/java.base/java/lang/Object.html" class="type-name-link external-link" title="class or interface in java.lang">Object</a>
 <ul>
-<li class="circle">Duke.<a href="Duke.html" class="type-name-link" title="class in Duke">Duke</a></li>
-<li class="circle">Duke.<a href="Parser.html" class="type-name-link" title="class in Duke">Parser</a></li>
-<li class="circle">Duke.<a href="Storage.html" class="type-name-link" title="class in Duke">Storage</a></li>
-<li class="circle">Duke.Task
+<li class="circle">Shirmin.<a href="Shirmin.html" class="type-name-link" title="class in Shirmin">Shirmin</a></li>
+<li class="circle">Shirmin.<a href="Parser.html" class="type-name-link" title="class in Shirmin">Parser</a></li>
+<li class="circle">Shirmin.<a href="Storage.html" class="type-name-link" title="class in Shirmin">Storage</a></li>
+<li class="circle">Shirmin.Task
 <ul>
-<li class="circle">Duke.<a href="Deadline.html" class="type-name-link" title="class in Duke">Deadline</a></li>
-<li class="circle">Duke.<a href="Event.html" class="type-name-link" title="class in Duke">Event</a></li>
-<li class="circle">Duke.<a href="Todo.html" class="type-name-link" title="class in Duke">Todo</a></li>
+<li class="circle">Shirmin.<a href="Deadline.html" class="type-name-link" title="class in Shirmin">Deadline</a></li>
+<li class="circle">Shirmin.<a href="Event.html" class="type-name-link" title="class in Shirmin">Event</a></li>
+<li class="circle">Shirmin.<a href="Todo.html" class="type-name-link" title="class in Shirmin">Todo</a></li>
 </ul>
 </li>
 </ul>
