+++ docs/allclasses-index.html
@@ -28,7 +28,7 @@ loadScripts(document, 'script');</script>
 <div class="top-nav" id="navbar-top"><button id="navbar-toggle-button" aria-controls="navbar-top" aria-expanded="false" aria-label="Toggle navigation links"><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span></button>
 <div class="skip-nav"><a href="#skip-navbar-top" title="Skip navigation links">Skip navigation links</a></div>
 <ul id="navbar-top-firstrow" class="nav-list" title="Navigation">
-<li><a href="Duke/package-summary.html">Package</a></li>
+<li><a href="Shirmin/package-summary.html">Package</a></li>
 <li>Class</li>
 <li><a href="overview-tree.html">Tree</a></li>
 <li><a href="index-files/index-1.html">Index</a></li>
@@ -55,27 +55,27 @@ loadScripts(document, 'script');</script>
 <div class="summary-table two-column-summary">
 <div class="table-header col-first">Class</div>
 <div class="table-header col-last">Description</div>
-<div class="col-first even-row-color all-classes-table all-classes-table-tab2"><a href="Duke/Deadline.html" title="class in Duke">Deadline</a></div>
+<div class="col-first even-row-color all-classes-table all-classes-table-tab2"><a href="Shirmin/Deadline.html" title="class in Shirmin">Deadline</a></div>
 <div class="col-last even-row-color all-classes-table all-classes-table-tab2">
 <div class="block">Represents a task with a deadline.</div>
 </div>
-<div class="col-first odd-row-color all-classes-table all-classes-table-tab2"><a href="Duke/Duke.html" title="class in Duke">Duke</a></div>
+<div class="col-first odd-row-color all-classes-table all-classes-table-tab2"><a href="Shirmin/Shirmin.html" title="class in Shirmin">Shirmin</a></div>
 <div class="col-last odd-row-color all-classes-table all-classes-table-tab2">
-<div class="block">Represents the main class of the Duke application.</div>
+<div class="block">Represents the main class of the Shirmin application.</div>
 </div>
-<div class="col-first even-row-color all-classes-table all-classes-table-tab2"><a href="Duke/Event.html" title="class in Duke">Event</a></div>
+<div class="col-first even-row-color all-classes-table all-classes-table-tab2"><a href="Shirmin/Event.html" title="class in Shirmin">Event</a></div>
 <div class="col-last even-row-color all-classes-table all-classes-table-tab2">
 <div class="block">Represents an event, represented by a description string and two LocalDateTime objects - from and to</div>
 </div>
-<div class="col-first odd-row-color all-classes-table all-classes-table-tab2"><a href="Duke/Parser.html" title="class in Duke">Parser</a></div>
+<div class="col-first odd-row-color all-classes-table all-classes-table-tab2"><a href="Shirmin/Parser.html" title="class in Shirmin">Parser</a></div>
 <div class="col-last odd-row-color all-classes-table all-classes-table-tab2">
 <div class="block">Represents a parser for interpreting user input and file data into Tasks.</div>
 </div>
-<div class="col-first even-row-color all-classes-table all-classes-table-tab2"><a href="Duke/Storage.html" title="class in Duke">Storage</a></div>
+<div class="col-first even-row-color all-classes-table all-classes-table-tab2"><a href="Shirmin/Storage.html" title="class in Shirmin">Storage</a></div>
 <div class="col-last even-row-color all-classes-table all-classes-table-tab2">
-<div class="block">Handles storage operations for Duke application tasks.</div>
+<div class="block">Handles storage operations for Shirmin application tasks.</div>
 </div>
-<div class="col-first odd-row-color all-classes-table all-classes-table-tab2"><a href="Duke/Todo.html" title="class in Duke">Todo</a></div>
+<div class="col-first odd-row-color all-classes-table all-classes-table-tab2"><a href="Shirmin/Todo.html" title="class in Shirmin">Todo</a></div>
 <div class="col-last odd-row-color all-classes-table all-classes-table-tab2">
 <div class="block">Represents a basic task without any additional date/time.</div>
 </div>
