+++ docs/allpackages-index.html
@@ -28,7 +28,7 @@ loadScripts(document, 'script');</script>
 <div class="top-nav" id="navbar-top"><button id="navbar-toggle-button" aria-controls="navbar-top" aria-expanded="false" aria-label="Toggle navigation links"><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span></button>
 <div class="skip-nav"><a href="#skip-navbar-top" title="Skip navigation links">Skip navigation links</a></div>
 <ul id="navbar-top-firstrow" class="nav-list" title="Navigation">
-<li><a href="Duke/package-summary.html">Package</a></li>
+<li><a href="Shirmin/package-summary.html">Package</a></li>
 <li>Class</li>
 <li><a href="overview-tree.html">Tree</a></li>
 <li><a href="index-files/index-1.html">Index</a></li>
@@ -54,7 +54,7 @@ loadScripts(document, 'script');</script>
 <div class="summary-table two-column-summary">
 <div class="table-header col-first">Package</div>
 <div class="table-header col-last">Description</div>
-<div class="col-first even-row-color"><a href="Duke/package-summary.html">Duke</a></div>
+<div class="col-first even-row-color"><a href="Shirmin/package-summary.html">Shirmin</a></div>
 <div class="col-last even-row-color">&nbsp;</div>
 </div>
 </main>
