+++ docs/index-files/index-10.html
@@ -28,7 +28,7 @@ loadScripts(document, 'script');</script>
 <div class="top-nav" id="navbar-top"><button id="navbar-toggle-button" aria-controls="navbar-top" aria-expanded="false" aria-label="Toggle navigation links"><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span></button>
 <div class="skip-nav"><a href="#skip-navbar-top" title="Skip navigation links">Skip navigation links</a></div>
 <ul id="navbar-top-firstrow" class="nav-list" title="Navigation">
-<li><a href="../Duke/package-summary.html">Package</a></li>
+<li><a href="../Shirmin/package-summary.html">Package</a></li>
 <li>Class</li>
 <li><a href="../overview-tree.html">Tree</a></li>
 <li class="nav-bar-cell1-rev">Index</li>
@@ -53,19 +53,19 @@ loadScripts(document, 'script');</script>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">D</a>&nbsp;<a href="index-3.html">E</a>&nbsp;<a href="index-4.html">F</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">M</a>&nbsp;<a href="index-8.html">P</a>&nbsp;<a href="index-9.html">R</a>&nbsp;<a href="index-10.html">S</a>&nbsp;<a href="index-11.html">T</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a>
 <h2 class="title" id="I:S">S</h2>
 <dl class="index">
-<dt><a href="../Duke/Storage.html#saveAllTasksToFile(java.util.ArrayList)" class="member-name-link">saveAllTasksToFile(ArrayList&lt;Task&gt;)</a> - Method in class Duke.<a href="../Duke/Storage.html" title="class in Duke">Storage</a></dt>
+<dt><a href="../Shirmin/Storage.html#saveAllTasksToFile(java.util.ArrayList)" class="member-name-link">saveAllTasksToFile(ArrayList&lt;Task&gt;)</a> - Method in class Shirmin.<a href="../Shirmin/Storage.html" title="class in Shirmin">Storage</a></dt>
 <dd>
 <div class="block">Saves a list of Task objects to the storage file.</div>
 </dd>
-<dt><a href="../Duke/Storage.html#saveTaskToFile(Duke.Task)" class="member-name-link">saveTaskToFile(Task)</a> - Method in class Duke.<a href="../Duke/Storage.html" title="class in Duke">Storage</a></dt>
+<dt><a href="../Shirmin/Storage.html#saveTaskToFile(Shirmin.Task)" class="member-name-link">saveTaskToFile(Task)</a> - Method in class Shirmin.<a href="../Shirmin/Storage.html" title="class in Shirmin">Storage</a></dt>
 <dd>
 <div class="block">Saves a single Task object to the storage file.</div>
 </dd>
-<dt><a href="../Duke/Storage.html" class="type-name-link" title="class in Duke">Storage</a> - Class in <a href="../Duke/package-summary.html">Duke</a></dt>
+<dt><a href="../Shirmin/Storage.html" class="type-name-link" title="class in Shirmin">Storage</a> - Class in <a href="../Shirmin/package-summary.html">Shirmin</a></dt>
 <dd>
-<div class="block">Handles storage operations for Duke application tasks.</div>
+<div class="block">Handles storage operations for Shirmin application tasks.</div>
 </dd>
-<dt><a href="../Duke/Storage.html#%3Cinit%3E(Duke.Ui,Duke.Parser)" class="member-name-link">Storage(Ui, Parser)</a> - Constructor for class Duke.<a href="../Duke/Storage.html" title="class in Duke">Storage</a></dt>
+<dt><a href="../Shirmin/Storage.html#%3Cinit%3E(Shirmin.Ui,Shirmin.Parser)" class="member-name-link">Storage(Ui, Parser)</a> - Constructor for class Shirmin.<a href="../Shirmin/Storage.html" title="class in Shirmin">Storage</a></dt>
 <dd>
 <div class="block">Initializes a new Storage object.</div>
 </dd>
