+++ docs/index-files/index-11.html
@@ -28,7 +28,7 @@ loadScripts(document, 'script');</script>
 <div class="top-nav" id="navbar-top"><button id="navbar-toggle-button" aria-controls="navbar-top" aria-expanded="false" aria-label="Toggle navigation links"><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span></button>
 <div class="skip-nav"><a href="#skip-navbar-top" title="Skip navigation links">Skip navigation links</a></div>
 <ul id="navbar-top-firstrow" class="nav-list" title="Navigation">
-<li><a href="../Duke/package-summary.html">Package</a></li>
+<li><a href="../Shirmin/package-summary.html">Package</a></li>
 <li>Class</li>
 <li><a href="../overview-tree.html">Tree</a></li>
 <li class="nav-bar-cell1-rev">Index</li>
@@ -53,25 +53,25 @@ loadScripts(document, 'script');</script>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">D</a>&nbsp;<a href="index-3.html">E</a>&nbsp;<a href="index-4.html">F</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">M</a>&nbsp;<a href="index-8.html">P</a>&nbsp;<a href="index-9.html">R</a>&nbsp;<a href="index-10.html">S</a>&nbsp;<a href="index-11.html">T</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a>
 <h2 class="title" id="I:T">T</h2>
 <dl class="index">
-<dt><a href="../Duke/Event.html#to" class="member-name-link">to</a> - Variable in class Duke.<a href="../Duke/Event.html" title="class in Duke">Event</a></dt>
+<dt><a href="../Shirmin/Event.html#to" class="member-name-link">to</a> - Variable in class Shirmin.<a href="../Shirmin/Event.html" title="class in Shirmin">Event</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../Duke/Todo.html" class="type-name-link" title="class in Duke">Todo</a> - Class in <a href="../Duke/package-summary.html">Duke</a></dt>
+<dt><a href="../Shirmin/Todo.html" class="type-name-link" title="class in Shirmin">Todo</a> - Class in <a href="../Shirmin/package-summary.html">Shirmin</a></dt>
 <dd>
 <div class="block">Represents a basic task without any additional date/time.</div>
 </dd>
-<dt><a href="../Duke/Todo.html#%3Cinit%3E(java.lang.String)" class="member-name-link">Todo(String)</a> - Constructor for class Duke.<a href="../Duke/Todo.html" title="class in Duke">Todo</a></dt>
+<dt><a href="../Shirmin/Todo.html#%3Cinit%3E(java.lang.String)" class="member-name-link">Todo(String)</a> - Constructor for class Shirmin.<a href="../Shirmin/Todo.html" title="class in Shirmin">Todo</a></dt>
 <dd>
 <div class="block">Initializes a new Todo object with a description.</div>
 </dd>
-<dt><a href="../Duke/Deadline.html#toString()" class="member-name-link">toString()</a> - Method in class Duke.<a href="../Duke/Deadline.html" title="class in Duke">Deadline</a></dt>
+<dt><a href="../Shirmin/Deadline.html#toString()" class="member-name-link">toString()</a> - Method in class Shirmin.<a href="../Shirmin/Deadline.html" title="class in Shirmin">Deadline</a></dt>
 <dd>
 <div class="block">Returns a string representation of the Deadline task.</div>
 </dd>
-<dt><a href="../Duke/Event.html#toString()" class="member-name-link">toString()</a> - Method in class Duke.<a href="../Duke/Event.html" title="class in Duke">Event</a></dt>
+<dt><a href="../Shirmin/Event.html#toString()" class="member-name-link">toString()</a> - Method in class Shirmin.<a href="../Shirmin/Event.html" title="class in Shirmin">Event</a></dt>
 <dd>
 <div class="block">Returns a string representation of the Event.</div>
 </dd>
-<dt><a href="../Duke/Todo.html#toString()" class="member-name-link">toString()</a> - Method in class Duke.<a href="../Duke/Todo.html" title="class in Duke">Todo</a></dt>
+<dt><a href="../Shirmin/Todo.html#toString()" class="member-name-link">toString()</a> - Method in class Shirmin.<a href="../Shirmin/Todo.html" title="class in Shirmin">Todo</a></dt>
 <dd>
 <div class="block">Returns a string representation of the Todo task.</div>
 </dd>
