+++ docs/index-files/index-5.html
@@ -28,7 +28,7 @@ loadScripts(document, 'script');</script>
 <div class="top-nav" id="navbar-top"><button id="navbar-toggle-button" aria-controls="navbar-top" aria-expanded="false" aria-label="Toggle navigation links"><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span></button>
 <div class="skip-nav"><a href="#skip-navbar-top" title="Skip navigation links">Skip navigation links</a></div>
 <ul id="navbar-top-firstrow" class="nav-list" title="Navigation">
-<li><a href="../Duke/package-summary.html">Package</a></li>
+<li><a href="../Shirmin/package-summary.html">Package</a></li>
 <li>Class</li>
 <li><a href="../overview-tree.html">Tree</a></li>
 <li class="nav-bar-cell1-rev">Index</li>
@@ -53,15 +53,15 @@ loadScripts(document, 'script');</script>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">D</a>&nbsp;<a href="index-3.html">E</a>&nbsp;<a href="index-4.html">F</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">M</a>&nbsp;<a href="index-8.html">P</a>&nbsp;<a href="index-9.html">R</a>&nbsp;<a href="index-10.html">S</a>&nbsp;<a href="index-11.html">T</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a>
 <h2 class="title" id="I:G">G</h2>
 <dl class="index">
-<dt><a href="../Duke/Deadline.html#getBy()" class="member-name-link">getBy()</a> - Method in class Duke.<a href="../Duke/Deadline.html" title="class in Duke">Deadline</a></dt>
+<dt><a href="../Shirmin/Deadline.html#getBy()" class="member-name-link">getBy()</a> - Method in class Shirmin.<a href="../Shirmin/Deadline.html" title="class in Shirmin">Deadline</a></dt>
 <dd>
 <div class="block">Returns the deadline time formatted as a string.</div>
 </dd>
-<dt><a href="../Duke/Event.html#getFrom()" class="member-name-link">getFrom()</a> - Method in class Duke.<a href="../Duke/Event.html" title="class in Duke">Event</a></dt>
+<dt><a href="../Shirmin/Event.html#getFrom()" class="member-name-link">getFrom()</a> - Method in class Shirmin.<a href="../Shirmin/Event.html" title="class in Shirmin">Event</a></dt>
 <dd>
 <div class="block">Returns a string representation of the start time of the event.</div>
 </dd>
-<dt><a href="../Duke/Event.html#getTo()" class="member-name-link">getTo()</a> - Method in class Duke.<a href="../Duke/Event.html" title="class in Duke">Event</a></dt>
+<dt><a href="../Shirmin/Event.html#getTo()" class="member-name-link">getTo()</a> - Method in class Shirmin.<a href="../Shirmin/Event.html" title="class in Shirmin">Event</a></dt>
 <dd>
 <div class="block">Returns a string representation of the end time of the event.</div>
 </dd>
