+++ docs/index-files/index-6.html
@@ -28,7 +28,7 @@ loadScripts(document, 'script');</script>
 <div class="top-nav" id="navbar-top"><button id="navbar-toggle-button" aria-controls="navbar-top" aria-expanded="false" aria-label="Toggle navigation links"><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span></button>
 <div class="skip-nav"><a href="#skip-navbar-top" title="Skip navigation links">Skip navigation links</a></div>
 <ul id="navbar-top-firstrow" class="nav-list" title="Navigation">
-<li><a href="../Duke/package-summary.html">Package</a></li>
+<li><a href="../Shirmin/package-summary.html">Package</a></li>
 <li>Class</li>
 <li><a href="../overview-tree.html">Tree</a></li>
 <li class="nav-bar-cell1-rev">Index</li>
@@ -53,11 +53,11 @@ loadScripts(document, 'script');</script>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">D</a>&nbsp;<a href="index-3.html">E</a>&nbsp;<a href="index-4.html">F</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">M</a>&nbsp;<a href="index-8.html">P</a>&nbsp;<a href="index-9.html">R</a>&nbsp;<a href="index-10.html">S</a>&nbsp;<a href="index-11.html">T</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a>
 <h2 class="title" id="I:H">H</h2>
 <dl class="index">
-<dt><a href="../Duke/Deadline.html#hasValidDate()" class="member-name-link">hasValidDate()</a> - Method in class Duke.<a href="../Duke/Deadline.html" title="class in Duke">Deadline</a></dt>
+<dt><a href="../Shirmin/Deadline.html#hasValidDate()" class="member-name-link">hasValidDate()</a> - Method in class Shirmin.<a href="../Shirmin/Deadline.html" title="class in Shirmin">Deadline</a></dt>
 <dd>
 <div class="block">Checks if the deadline has a valid date set.</div>
 </dd>
-<dt><a href="../Duke/Event.html#hasValidDates()" class="member-name-link">hasValidDates()</a> - Method in class Duke.<a href="../Duke/Event.html" title="class in Duke">Event</a></dt>
+<dt><a href="../Shirmin/Event.html#hasValidDates()" class="member-name-link">hasValidDates()</a> - Method in class Shirmin.<a href="../Shirmin/Event.html" title="class in Shirmin">Event</a></dt>
 <dd>
 <div class="block">Checks if both start date and end date are set for an event.</div>
 </dd>
