+++ docs/index.html
@@ -8,11 +8,11 @@
 <meta name="dc.created" content="2024-01-30">
 <meta name="description" content="index redirect">
 <meta name="generator" content="javadoc/IndexRedirectWriter">
-<link rel="canonical" href="Duke/package-summary.html">
+<link rel="canonical" href="Shirmin/package-summary.html">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
-<script type="text/javascript">window.location.replace('Duke/package-summary.html')</script>
+<script type="text/javascript">window.location.replace('Shirmin/package-summary.html')</script>
 <noscript>
-<meta http-equiv="Refresh" content="0;Duke/package-summary.html">
+<meta http-equiv="Refresh" content="0;Shirmin/package-summary.html">
 </noscript>
 </head>
 <body class="index-redirect-page">
@@ -20,7 +20,7 @@
 <noscript>
 <p>JavaScript is disabled on your browser.</p>
 </noscript>
-<p><a href="Duke/package-summary.html">Duke/package-summary.html</a></p>
+<p><a href="Shirmin/package-summary.html">Shirmin/package-summary.html</a></p>
 </main>
 </body>
 </html>
