+++ docs/overview-tree.html
@@ -51,7 +51,7 @@ loadScripts(document, 'script');</script>
 <h1 class="title">Hierarchy For All Packages</h1>
 <span class="package-hierarchy-label">Package Hierarchies:</span>
 <ul class="horizontal">
-<li><a href="Duke/package-tree.html">Duke</a></li>
+<li><a href="Shirmin/package-tree.html">Shirmin</a></li>
 </ul>
 </div>
 <section class="hierarchy">
@@ -59,14 +59,14 @@ loadScripts(document, 'script');</script>
 <ul>
 <li class="circle">java.lang.<a href="https://docs.oracle.com/en/java/javase/19/docs/api/java.base/java/lang/Object.html" class="type-name-link external-link" title="class or interface in java.lang">Object</a>
 <ul>
-<li class="circle">Duke.<a href="Duke/Duke.html" class="type-name-link" title="class in Duke">Duke</a></li>
-<li class="circle">Duke.<a href="Duke/Parser.html" class="type-name-link" title="class in Duke">Parser</a></li>
-<li class="circle">Duke.<a href="Duke/Storage.html" class="type-name-link" title="class in Duke">Storage</a></li>
-<li class="circle">Duke.Task
+<li class="circle">Shirmin.<a href="Shirmin/Shirmin.html" class="type-name-link" title="class in Shirmin">Shirmin</a></li>
+<li class="circle">Shirmin.<a href="Shirmin/Parser.html" class="type-name-link" title="class in Shirmin">Parser</a></li>
+<li class="circle">Shirmin.<a href="Shirmin/Storage.html" class="type-name-link" title="class in Shirmin">Storage</a></li>
+<li class="circle">Shirmin.Task
 <ul>
-<li class="circle">Duke.<a href="Duke/Deadline.html" class="type-name-link" title="class in Duke">Deadline</a></li>
-<li class="circle">Duke.<a href="Duke/Event.html" class="type-name-link" title="class in Duke">Event</a></li>
-<li class="circle">Duke.<a href="Duke/Todo.html" class="type-name-link" title="class in Duke">Todo</a></li>
+<li class="circle">Shirmin.<a href="Shirmin/Deadline.html" class="type-name-link" title="class in Shirmin">Deadline</a></li>
+<li class="circle">Shirmin.<a href="Shirmin/Event.html" class="type-name-link" title="class in Shirmin">Event</a></li>
+<li class="circle">Shirmin.<a href="Shirmin/Todo.html" class="type-name-link" title="class in Shirmin">Todo</a></li>
 </ul>
 </li>
 </ul>
