+++ src/main/java/Shirmin/Deadline.java
@@ -1,4 +1,4 @@
-package Duke;
+package Shirmin;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
