+++ src/test/java/Shirmin/EventTest.java
@@ -1,4 +1,4 @@
-package Duke;
+package Shirmin;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
