+++ text-ui-test/runtest.bat
@@ -9,7 +9,7 @@ if exist ACTUAL.TXT del ACTUAL.TXT
 REM compile the code into the bin folder
 
 
-javac -cp "..\src\main\java\Duke\Duke.java" -Xlint:none -d ..\bin "..\src\main\java\Duke\*.java"
+javac -cp "..\src\main\java\Shirmin\Shirmin.java" -Xlint:none -d ..\bin "..\src\main\java\Shirmin\*.java"
 
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
@@ -18,7 +18,7 @@ IF ERRORLEVEL 1 (
 REM no error here, errorlevel == 0
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin Duke< input.txt > ACTUAL.TXT
+java -classpath ..\bin Shirmin< input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC ACTUAL.TXT EXPECTED.TXT
